+++ src/com/xqmsg/sdk/v2/services/dashboard/RemoveApplication.ts
@@ -5,38 +5,36 @@ import XQModule from "../XQModule";
 import XQSDK from "../../XQSDK";
 
 /**
- * A service which is utilized to remove an existing Contact
+ * A service which is utilized to remove an existing developer application.
  *
- * @class [RemoveContact]
+ * @class [RemoveApplication]
  */
-export default class RemoveContact extends XQModule {
+export default class RemoveApplication extends XQModule {
   /** The required fields of the payload needed to utilize the service */
   requiredFields: string[];
 
   /** Specified name of the service */
   serviceName: string;
 
-  /** The field name representing the id */
+  /** The field name representing the ID of the application */
   static ID: "id" = "id";
 
   /**
    * @param {Map} maybePayLoad - Container for the request parameters supplied to this method.
-   * @param {String} id - the user id of the Contact that will be removed
+   * @param {Number} id - the id of the application
    * @returns {Promise<ServerResponse<{payload:{}}>>}
    */
   supplyAsync: (maybePayload: {
-    [RemoveContact.ID]: string;
+    [RemoveApplication.ID]: number;
   }) => Promise<ServerResponse>;
 
   constructor(sdk: XQSDK) {
     super(sdk);
-    this.serviceName = "contact";
-    this.requiredFields = [RemoveContact.ID];
+    this.serviceName = "devapp";
+    this.requiredFields = [RemoveApplication.ID];
 
     this.supplyAsync = (maybePayLoad) => {
       try {
-        this.sdk.validateInput(maybePayLoad, this.requiredFields);
-
         const dashboardAccessToken = this.sdk.validateAccessToken(
           Destination.DASHBOARD
         );
@@ -47,7 +45,7 @@ export default class RemoveContact extends XQModule {
 
         return this.sdk.call(
           this.sdk.DASHBOARD_SERVER_URL,
-          `${this.serviceName}/${maybePayLoad[RemoveContact.ID]}?delete=true`,
+          this.serviceName + "/" + maybePayLoad[RemoveApplication.ID],
           CallMethod.DELETE,
           additionalHeaderProperties,
           null,