+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.6.0] - 2022-04-18
+
+### Addition
+- Add CRUD operations for businesses, including `AddBusiness`, `GetCurrentBusiness`, and `UpdateBusiness`.
+
+- Add CRUD operations for applications, including `AddApplication`, `RemoveApplication`, and `UpdateApplication`.
+### Modified
+
+- Update the `README.md` documentation file to reflect recent changes to the sdk.
+
 ## [1.5.3] - 2022-04-18
 ### Modified